requirements/requirements.txt
Babel==1.3
Flask==0.10.1
Flask-Admin==1.0.8
Flask-Assets==0.10
Flask-Babel==0.9
Flask-Login==0.2.11
Flask-Mail==0.9.1
Flask-Migrate==1.3.0
Flask-SQLAlchemy==2.0
Flask-Script==2.0.5
Flask-SocketIO==0.3.8
Flask-WTF==0.10.3
Jinja2==2.7.3
Mako==1.0.0
MarkupSafe==0.23
SQLAlchemy==0.9.8
Tempita==0.5.2
WTForms==2.0.1
Werkzeug==0.9.6
alembic==0.7.3
amqp==1.4.6
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.19
blinker==1.3
cssmin==0.2.0
decorator==3.4.0
gevent==1.0.1
gevent-socketio==0.3.6
gevent-websocket==0.9.3
greenlet==0.4.4
httplib2==0.9
itsdangerous==0.24
kombu==3.0.24
nose==1.3.4
pbr==0.10.0
pyserial==2.7
python-rrdtool==1.4.7
pytz==2014.7
recordtype==1.1
six==1.8.0
speaklater==1.3
sqlalchemy-migrate==0.9.2
sqlparse==0.1.13
twilio==3.6.9
tzlocal==1.1.2
whisper==0.9.12
wsgiref==0.1.2
